+++ editor/src/editor-extension.ts
@@ -1,0 +1,10 @@
+import { ValueEditorFactory } from './value-editors';
+
+export interface EditorExtension {
+	valueEditors?: ValueEditorExtension[];
+}
+
+export interface ValueEditorExtension {
+	editorId: string;
+	factory: ValueEditorFactory;
+}