+++ model/src/value-models/boolean/boolean-value-model-configuration.ts
@@ -1,3 +1,4 @@
 export interface BooleanValueModelConfiguration {
 	defaultValue?: boolean;
+	editorId?: string;
 }