+++ model/src/value-models/boolean/boolean-value-model.ts
@@ -1,6 +1,7 @@
 import { ValueModel, ValueModelFactoryFromModel } from '../../model';
 import { Path } from '../../core/path';
 import { BooleanValueModelConfiguration } from './boolean-value-model-configuration';
+import { booleanValueModelValidator } from './boolean-value-model-validator';
 
 export type BooleanValueModel = ValueModel<boolean, BooleanValueModelConfiguration>;
 
@@ -9,6 +10,7 @@ export const booleanValueModelId = 'boolean';
 export const createBooleanValueModel = (configuration: BooleanValueModelConfiguration): ValueModelFactoryFromModel<BooleanValueModel> => ({
 	create: (path: Path) => ({
 		id: booleanValueModelId,
+		editorId: configuration.editorId,
 		label: 'Boolean',
 		path,
 		configuration,
@@ -19,6 +21,6 @@ export const createBooleanValueModel = (configuration: BooleanValueModelConfigur
 			return false;
 		},
 		getVariableDefinitions: () => null,
-		validate: () => null
+		validate: booleanValueModelValidator
 	})
 });