+++ model/src/value-models/branches/index.ts
@@ -1,1 +1,2 @@
+export * from './branches-value-model-configuration';
 export * from './branches-value-model';