+++ model/src/value-models/number/number-value-model-configuration.ts
@@ -2,4 +2,5 @@ export interface NumberValueModelConfiguration {
 	defaultValue?: number;
 	min?: number;
 	max?: number;
+	editorId?: string;
 }