+++ model/src/value-models/string/string-value-model-configuration.ts
@@ -3,4 +3,5 @@ export interface StringValueModelConfiguration {
 	defaultValue?: string;
 	pattern?: RegExp;
 	multiline?: boolean | number;
+	editorId?: string;
 }